requirements.txt
requests
beautifulsoup4
lxml
PyYAML
gspread
oauth2client
python-dotenv